backend/requirements.txt
# API
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
cassandra-driver>=3.25.0  # This is the correct package for ScyllaDB/Cassandra
pandas>=1.3.0

# ML
scikit-surprise>=1.1.1

# Utils
python-dotenv>=0.19.0
pydantic>=1.8.0
